Add Nix flake
This is another take on #1026 that aims to minimally implement a Nix development environment.

Notable differences are:

- Uses `pyproject.nix` for auto-packaging of the environment
  - Mach-nix is unmaintained
    Not only is it unmaintained but extremely heavy weight.

  - Uses nixpkgs Python packages
    Which means better binary cache performance than mach-nix offers.

- No direnv additions
I use direnv myself but expect users to be able to `echo 'use flake' > .envrc` if they want to use direnv.

- Uses zephyr sources from a flake input
This should make maintenance simpler as there is no manual updating required.